engines/io_uring: io_uring_cmd engine fixes
The io_uring_cmd ioengine assumes that logical block size is always
power of 2. But with namespace formats where metadata is transferred
at the end of logical block i.e. an extended logical block this is
not true.
This patch calculates the correct extended logical block size and
uses division operation for start and number of logical block
calculation. The existing calculation for power of 2 logical block
size will remain same.
This also add checks to verify that block sizes are multiple of LBA size.

This current implementation however doesn't support protection info
and metadata transferred as separate buffer. Return error for those.

Signed-off-by: Ankit Kumar <ankit.kumar@samsung.com>