os-linux: add initial support for IO priority hints
Add initial support for Linux to allow specifying a hint for any
priority value. With this change, a priority value becomes the
combination of a priority class, a priority level and a hint.
The generic os.h ioprio manipulation macros, as well as the
os-dragonfly.h ioprio manipulation macros  are modified to ignore this
hint.

Signed-off-by: Damien Le Moal <damien.lemoal@opensource.wdc.com>